SCCModular: Consolidated modular SCC components into fewer modules
