+++ transformations/transformations/__init__.py
@@ -10,14 +10,10 @@ from transformations.argument_shape import * # noqa
 from transformations.data_offload import * # noqa
 from transformations.single_column_claw import * # noqa
 from transformations.single_column_coalesced import * # noqa
+from transformations.single_column_coalesced_wrapper import * # noqa
+from transformations.single_column_coalesced_vector import * # noqa
 from transformations.utility_routines import * # noqa
 from transformations.scc_cuf import * # noqa
 from transformations.pool_allocator import * # noqa
-from transformations.scc_base import * # noqa
-from transformations.scc_devector import * # noqa
-from transformations.scc_demote import * # noqa
-from transformations.scc_revector import * # noqa
-from transformations.scc_hoist import * # noqa
-from transformations.scc_annotate import * # noqa
 
 __version__ = "0.0.0"
