+++ transformations/transformations/single_column_coalesced_vector.py
@@ -1,0 +1,365 @@
+# (C) Copyright 2018- ECMWF.
+# This software is licensed under the terms of the Apache Licence Version 2.0
+# which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
+# In applying this licence, ECMWF does not waive the privileges and immunities
+# granted to it by virtue of its status as an intergovernmental organisation
+# nor does it submit to any jurisdiction.
+
+from more_itertools import split_at
+
+from loki.expression import symbols as sym
+from loki import (
+     Transformation, FindNodes, ir, FindScopes, as_tuple, flatten, Transformer,
+     NestedTransformer, FindVariables, demote_variables, is_dimension_constant
+)
+from transformations.single_column_coalesced import SCCBaseTransformation
+
+__all__ = ['SCCDevectorTransformation', 'SCCRevectorTransformation', 'SCCDemoteTransformation']
+
+
+class SCCDevectorTransformation(Transformation):
+    """
+    A set of utilities that can be used to strip vector loops from a :any:`Subroutine`
+    and determine the regions of the IR to be placed within thread-parallel loop directives.
+
+    Parameters
+    ----------
+    horizontal : :any:`Dimension`
+        :any:`Dimension` object describing the variable conventions used in code
+        to define the horizontal data dimension and iteration space.
+    """
+
+    def __init__(self, horizontal):
+        self.horizontal = horizontal
+
+        self._processed = {}
+
+    @classmethod
+    def kernel_remove_vector_loops(cls, routine, horizontal):
+        """
+        Remove all vector loops over the specified dimension.
+
+        Parameters
+        ----------
+        routine : :any:`Subroutine`
+            The subroutine in the vector loops should be removed.
+        horizontal : :any:`Dimension`
+            The dimension specifying the horizontal vector dimension
+        """
+        loop_map = {}
+        for loop in FindNodes(ir.Loop).visit(routine.body):
+            if loop.variable == horizontal.index:
+                loop_map[loop] = loop.body
+        routine.body = Transformer(loop_map).visit(routine.body)
+
+    @classmethod
+    def extract_vector_sections(cls, section, horizontal):
+        """
+        Extract a contiguous sections of nodes that contains vector-level
+        computations and are not interrupted by recursive subroutine calls
+        or nested control-flow structures.
+
+        Parameters
+        ----------
+        section : tuple of :any:`Node`
+            A section of nodes from which to extract vector-level sub-sections
+        horizontal: :any:`Dimension`
+            The dimension specifying the horizontal vector dimension
+        """
+
+        _scope_note_types = (ir.Loop, ir.Conditional, ir.MultiConditional)
+
+        # Identify outer "scopes" (loops/conditionals) constrained by recursive routine calls
+        separator_nodes = []
+        calls = FindNodes(ir.CallStatement).visit(section)
+        for call in calls:
+            if call in section:
+                # If the call is at the current section's level, it's a separator
+                separator_nodes.append(call)
+
+            else:
+                # If the call is deeper in the IR tree, it's highest ancestor is used
+                ancestors = flatten(FindScopes(call).visit(section))
+                ancestor_scopes = [a for a in ancestors if isinstance(a, _scope_note_types)]
+                if len(ancestor_scopes) > 0 and ancestor_scopes[0] not in separator_nodes:
+                    separator_nodes.append(ancestor_scopes[0])
+
+        # Extract contiguous node sections between separator nodes
+        assert all(n in section for n in separator_nodes)
+        subsections = [as_tuple(s) for s in split_at(section, lambda n: n in separator_nodes)]
+
+        # Filter sub-sections that do not use the horizontal loop index variable
+        subsections = [s for s in subsections if horizontal.index in list(FindVariables().visit(s))]
+
+        # Recurse on all separator nodes that might contain further vector sections
+        for separator in separator_nodes:
+
+            if isinstance(separator, ir.Loop):
+                subsec_body = cls.extract_vector_sections(separator.body, horizontal)
+                if subsec_body:
+                    subsections += subsec_body
+
+            if isinstance(separator, ir.Conditional):
+                subsec_body = cls.extract_vector_sections(separator.body, horizontal)
+                if subsec_body:
+                    subsections += subsec_body
+                subsec_else = cls.extract_vector_sections(separator.else_body, horizontal)
+                if subsec_else:
+                    subsections += subsec_else
+
+            if isinstance(separator, ir.MultiConditional):
+                for body in separator.bodies:
+                    subsec_body = cls.extract_vector_sections(body, horizontal)
+                    if subsec_body:
+                        subsections += subsec_body
+                subsec_else = cls.extract_vector_sections(separator.else_body, horizontal)
+                if subsec_else:
+                    subsections += subsec_else
+
+        return subsections
+
+    def transform_subroutine(self, routine, **kwargs):
+        """
+        Apply SCCDevector utilities to a :any:`Subroutine`.
+
+        Parameters
+        ----------
+        routine : :any:`Subroutine`
+            Subroutine to apply this transformation to.
+        role : string
+            Role of the subroutine in the call tree; should be ``"kernel"``
+        """
+
+        # TODO: we only need this here until the scheduler can combine multiple transformations into single pass
+        # Bail if routine has already been processed
+        if self._processed.get(routine, None):
+            return
+
+        role = kwargs['role']
+
+        if role == 'kernel':
+            self.process_kernel(routine)
+
+        # Mark routine as processed
+        self._processed[routine] = True
+
+    def process_kernel(self, routine):
+        """
+        Applies the SCCDevector utilities to a "kernel". This consists simply
+        of stripping vector loops and determing which sections of the IR can be
+        placed within thread-parallel loops.
+
+        Parameters
+        ----------
+        routine : :any:`Subroutine`
+            Subroutine to apply this transformation to.
+        """
+
+        # Remove all vector loops over the specified dimension
+        self.kernel_remove_vector_loops(routine, self.horizontal)
+
+        # Replace sections with marked Section node
+        section_mapper = {s: ir.Section(body=s, label='vector_section')
+                          for s in self.extract_vector_sections(routine.body.body, self.horizontal)}
+        routine.body = NestedTransformer(section_mapper).visit(routine.body)
+
+
+class SCCRevectorTransformation(Transformation):
+    """
+    A transformation to wrap thread-parallel IR sections within a horizontal loop.
+    This transformation relies on markers placed by :any:`SCCDevectorTransformation`.
+
+    Parameters
+    ----------
+    horizontal : :any:`Dimension`
+        :any:`Dimension` object describing the variable conventions used in code
+        to define the horizontal data dimension and iteration space.
+    """
+
+    def __init__(self, horizontal):
+        self.horizontal = horizontal
+        self._processed = {}
+
+    @classmethod
+    def wrap_vector_section(cls, section, routine, horizontal):
+        """
+        Wrap a section of nodes in a vector-level loop across the horizontal.
+
+        Parameters
+        ----------
+        section : tuple of :any:`Node`
+            A section of nodes to be wrapped in a vector-level loop
+        routine : :any:`Subroutine`
+            The subroutine in the vector loops should be removed.
+        horizontal: :any:`Dimension`
+            The dimension specifying the horizontal vector dimension
+        """
+
+        # Create a single loop around the horizontal from a given body
+        v_start = routine.variable_map[horizontal.bounds[0]]
+        v_end = routine.variable_map[horizontal.bounds[1]]
+        index = SCCBaseTransformation.get_integer_variable(routine, horizontal.index)
+        bounds = sym.LoopRange((v_start, v_end))
+
+        # Ensure we clone all body nodes, to avoid recursion issues
+        vector_loop = ir.Loop(variable=index, bounds=bounds, body=Transformer().visit(section))
+
+        # Add a comment before the pragma-annotated loop to ensure
+        # we do not overlap with neighbouring pragmas
+        return (ir.Comment(''), vector_loop)
+
+    def transform_subroutine(self, routine, **kwargs):
+        """
+        Apply SCCRevector utilities to a :any:`Subroutine`.
+
+        Parameters
+        ----------
+        routine : :any:`Subroutine`
+            Subroutine to apply this transformation to.
+        role : string
+            Role of the subroutine in the call tree; should be ``"kernel"``
+        """
+
+        # TODO: we only need this here until the scheduler can combine multiple transformations into single pass
+        # Bail if routine has already been processed
+        if self._processed.get(routine, None):
+            return
+
+        role = kwargs['role']
+
+        if role == 'kernel':
+            self.process_kernel(routine)
+
+        # Mark routine as processed
+        self._processed[routine] = True
+
+    def process_kernel(self, routine):
+        """
+        Applies the SCCRevector utilities to a "kernel" and wraps all thread-parallel sections within
+        a horizontal loop. The markers placed by :any:`SCCDevectorTransformation` are removed.
+
+        Parameters
+        ----------
+        routine : :any:`Subroutine`
+            Subroutine to apply this transformation to.
+        """
+
+        # Promote vector loops to be the outermost loop dimension in the kernel
+        mapper = {s.body: self.wrap_vector_section(s.body, routine, self.horizontal)
+                          for s in FindNodes(ir.Section).visit(routine.body)
+                          if s.label == 'vector_section'}
+        routine.body = NestedTransformer(mapper).visit(routine.body)
+
+
+class SCCDemoteTransformation(Transformation):
+    """
+    A set of utilities to determine which local arrays can be safely demoted in a
+    :any:`Subroutine` as part of a transformation pass.
+
+    Parameters
+    ----------
+    horizontal : :any:`Dimension`
+        :any:`Dimension` object describing the variable conventions used in code
+        to define the horizontal data dimension and iteration space.
+    """
+
+    def __init__(self, horizontal, demote_local_arrays=True):
+        self.horizontal = horizontal
+
+        self.demote_local_arrays = demote_local_arrays
+        self._processed = {}
+
+    @classmethod
+    def kernel_get_locals_to_demote(cls, routine, sections, horizontal):
+
+        argument_names = [v.name for v in routine.arguments]
+
+        def _get_local_arrays(section):
+            """
+            Filters out local argument arrays that solely buffer the
+            horizontal vector dimension
+            """
+            arrays = FindVariables(unique=False).visit(section)
+            # Only demote local arrays with the horizontal as fast dimension
+            arrays = [v for v in arrays if isinstance(v, sym.Array)]
+            arrays = [v for v in arrays if v.name not in argument_names]
+            arrays = [v for v in arrays if v.shape and v.shape[0] == horizontal.size]
+
+            # Also demote arrays whose remaning dimensions are known constants
+            arrays = [v for v in arrays if all(is_dimension_constant(d) for d in v.shape[1:])]
+            return arrays
+
+        # Create a list of all local horizontal temporary arrays
+        candidates = _get_local_arrays(routine.body)
+
+        # Create an index into all variable uses per vector-level section
+        vars_per_section = {s: set(v.name.lower() for v in _get_local_arrays(s)) for s in sections}
+
+        # Count in how many sections each temporary is used
+        counts = {}
+        for arr in candidates:
+            counts[arr] = sum(1 if arr.name.lower() in v else 0 for v in vars_per_section.values())
+
+        # Mark temporaries that are only used in one section for demotion
+        to_demote = [k for k, v in counts.items() if v == 1]
+
+        # Filter out variables that we will pass down the call tree
+        calls = FindNodes(ir.CallStatement).visit(routine.body)
+        call_args = flatten(call.arguments for call in calls)
+        call_args += flatten(list(dict(call.kwarguments).values()) for call in calls)
+        to_demote = [v for v in to_demote if v.name not in call_args]
+
+        return set(to_demote)
+
+    def transform_subroutine(self, routine, **kwargs):
+        """
+        Apply SCCDemote utilities to a :any:`Subroutine`.
+
+        Parameters
+        ----------
+        routine : :any:`Subroutine`
+            Subroutine to apply this transformation to.
+        role : string
+            Role of the subroutine in the call tree; should be ``"kernel"``
+        """
+
+        # TODO: we only need this here until the scheduler can combine multiple transformations into single pass
+        # Bail if routine has already been processed
+        if self._processed.get(routine, None):
+            return
+
+        role = kwargs['role']
+        item = kwargs.get('item', None)
+
+        if role == 'kernel':
+            demote_locals = self.demote_local_arrays
+            if item:
+                demote_locals = item.config.get('demote_locals', self.demote_local_arrays)
+            self.process_kernel(routine, demote_locals=demote_locals)
+
+        # Mark routine as processed
+        self._processed[routine] = True
+
+    def process_kernel(self, routine, demote_locals=True):
+        """
+        Applies the SCCDemote utilities to a "kernel" and demotes all suitable local arrays.
+
+        Parameters
+        ----------
+        routine : :any:`Subroutine`
+            Subroutine to apply this transformation to.
+        """
+
+        # Find vector sections marked in the SCCDevectorTransformation
+        sections = [s for s in FindNodes(ir.Section).visit(routine.body) if s.label == 'vector_section']
+
+        # Extract the local variables to demote after we wrap the sections in vector loops.
+        # We do this, because need the section blocks to determine which local arrays
+        # may carry buffered values between them, so that we may not demote those!
+        to_demote = self.kernel_get_locals_to_demote(routine, sections, self.horizontal)
+
+        # Demote all private local variables that do not buffer values between sections
+        if demote_locals:
+            variables = tuple(v.name for v in to_demote)
+            if variables:
+                demote_variables(routine, variable_names=variables, dimensions=self.horizontal.size)
